Revert "fix lib-modal"
