+++ src/partials/empty-lib-modal.html
@@ -1,4 +1,4 @@
-<div data-modal-lib class="lib-backdrop lib-is-hidden">
+<div data-modal class="backdrop is-hidden">
   <div class="lib-modal">
     <button class="lib_modal-close-btn" data-modal-close>
       <svg class="lib_button-close__icon-close" width="10" height="10">
